backend_in_django/requirements.txt
django==4.1.7
psycopg2==2.9.5
djangorestframework==3.13.1
requests==2.28.1